rename wikivis -> wikiviz
