+++ backend/scrapy/scrapy.cfg
@@ -1,5 +1,5 @@
 [settings]
-default = wikivis.settings
+default = wikiviz.settings
 
 [deploy]
-project = wikivis
+project = wikiviz
