+++ backend/scrapy/wikiviz/utils/network/serializer.py
@@ -1,7 +1,7 @@
 import pickle
 from networkx import Graph
 
-from wikivis.utils.create_missing_folders import create_missing_folders
+from wikiviz.utils.create_missing_folders import create_missing_folders
 
 
 class Serializer:
